prog_game/requirements.txt
asgiref==3.9.2
certifi==2025.8.3
charset-normalizer==3.4.3
Django==5.2.7
django-cors-headers==4.9.0
django-restframework==0.0.1
djangorestframework==3.16.1
idna==3.10
psycopg2-binary==2.9.10
PyJWT==2.10.1
requests==2.32.5
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.5.0